Still troubleshooting.
Idk man, maybe my package is just busted. But not having access to the .sql file within my netbeans might be the problem.

I didn't get any errors when I typed in a new password, seems to have submitted but not done anything on the backend. This security demos lab was buggy as hell.